requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
openai>=1.14.0,<2.0.0
